tests stabilization by splitting long client test in smaller chunks
