+++ js/apps/admin-ui/cypress/e2e/clients_test.spec.ts
@@ -1,36 +1,24 @@
 import LoginPage from "../support/pages/LoginPage";
-import ListingPage, {
-  Filter,
-  FilterAssignedType,
-} from "../support/pages/admin-ui/ListingPage";
+import ListingPage from "../support/pages/admin-ui/ListingPage";
 import CreateClientPage from "../support/pages/admin-ui/manage/clients/CreateClientPage";
 import adminClient from "../support/util/AdminClient";
 import { keycloakBefore } from "../support/util/keycloak_hooks";
 import RoleMappingTab from "../support/pages/admin-ui/manage/RoleMappingTab";
-import createRealmRolePage from "../support/pages/admin-ui/manage/realm_roles/CreateRealmRolePage";
-import AssociatedRolesPage from "../support/pages/admin-ui/manage/realm_roles/AssociatedRolesPage";
 import ClientRolesTab from "../support/pages/admin-ui/manage/clients/ClientRolesTab";
-import InitialAccessTokenTab from "../support/pages/admin-ui/manage/clients/tabs/InitialAccessTokenTab";
-import AdvancedTab from "../support/pages/admin-ui/manage/clients/client_details/tabs/AdvancedTab";
 import ClientDetailsPage, {
   ClientsDetailsTab,
 } from "../support/pages/admin-ui/manage/clients/client_details/ClientDetailsPage";
 import CommonPage from "../support/pages/CommonPage";
-import AttributesTab from "../support/pages/admin-ui/manage/AttributesTab";
 import DedicatedScopesMappersTab from "../support/pages/admin-ui/manage/clients/client_details/DedicatedScopesMappersTab";
 import { ClientRegistrationPage } from "../support/pages/admin-ui/manage/clients/ClientRegistrationPage";
-import RealmSettingsPage from "../support/pages/admin-ui/manage/realm_settings/RealmSettingsPage";
 
 let itemId = "client_crud";
 const loginPage = new LoginPage();
-const associatedRolesPage = new AssociatedRolesPage();
 const createClientPage = new CreateClientPage();
 const clientDetailsPage = new ClientDetailsPage();
 const commonPage = new CommonPage();
 const listingPage = new ListingPage();
-const attributesTab = new AttributesTab();
 const dedicatedScopesMappersTab = new DedicatedScopesMappersTab();
-const realmSettings = new RealmSettingsPage();
 
 describe("Clients test", () => {
   const realmName = `clients-realm-${crypto.randomUUID()}`;
@@ -38,311 +26,6 @@ describe("Clients test", () => {
   before(() => adminClient.createRealm(realmName));
 
   after(() => adminClient.deleteRealm(realmName));
-
-  describe("Client details - Client scopes subtab", () => {
-    const clientId = "client-scopes-subtab-test";
-    const clientScopeName = "client-scope-test";
-    const clientScopeNameDefaultType = "client-scope-test-default-type";
-    const clientScopeNameOptionalType = "client-scope-test-optional-type";
-    const clientScope = {
-      name: clientScopeName,
-      description: "",
-      protocol: "openid-connect",
-      attributes: {
-        "include.in.token.scope": "true",
-        "display.on.consent.screen": "true",
-        "gui.order": "1",
-        "consent.screen.text": "",
-      },
-    };
-    const msgScopeMappingRemoved = "Scope mapping successfully removed";
-
-    before(async () => {
-      await adminClient.inRealm(realmName, () =>
-        adminClient.createClient({
-          clientId,
-          protocol: "openid-connect",
-          publicClient: false,
-        }),
-      );
-      for (let i = 0; i < 5; i++) {
-        clientScope.name = clientScopeName + i;
-        await adminClient.inRealm(realmName, () =>
-          adminClient.createClientScope(clientScope),
-        );
-        await adminClient.inRealm(realmName, () =>
-          adminClient.addDefaultClientScopeInClient(
-            clientScopeName + i,
-            clientId,
-          ),
-        );
-      }
-      clientScope.name = clientScopeNameDefaultType;
-      await adminClient.inRealm(realmName, () =>
-        adminClient.createClientScope(clientScope),
-      );
-      clientScope.name = clientScopeNameOptionalType;
-      await adminClient.inRealm(realmName, () =>
-        adminClient.createClientScope(clientScope),
-      );
-    });
-
-    beforeEach(() => {
-      loginPage.logIn();
-      keycloakBefore();
-      commonPage.sidebar().goToRealm(realmName);
-      commonPage.sidebar().goToClients();
-      commonPage.tableToolbarUtils().searchItem(clientId);
-      cy.intercept(`/admin/realms/${realmName}/clients/*`).as("fetchClient");
-      commonPage.tableUtils().clickRowItemLink(clientId);
-      cy.wait("@fetchClient");
-      clientDetailsPage.goToClientScopesTab();
-    });
-
-    after(async () => {
-      adminClient.inRealm(realmName, () => adminClient.deleteClient(clientId));
-      for (let i = 0; i < 5; i++) {
-        await adminClient.inRealm(realmName, () =>
-          adminClient.deleteClientScope(clientScopeName + i),
-        );
-      }
-      await adminClient.inRealm(realmName, () =>
-        adminClient.deleteClientScope(clientScopeNameDefaultType),
-      );
-      await adminClient.inRealm(realmName, () =>
-        adminClient.deleteClientScope(clientScopeNameOptionalType),
-      );
-    });
-
-    it("Should list client scopes", () => {
-      commonPage
-        .tableUtils()
-        .checkRowItemsGreaterThan(1)
-        .checkRowItemExists(clientScopeName + 0);
-    });
-
-    it("Should search existing client scope by name", () => {
-      commonPage.tableToolbarUtils().searchItem(clientScopeName + 0, false);
-      commonPage
-        .tableUtils()
-        .checkRowItemExists(clientScopeName + 0)
-        .checkRowItemsEqualTo(1);
-    });
-
-    it("Should search non-existent client scope by name", () => {
-      commonPage.tableToolbarUtils().searchItem("non-existent-item", false);
-      commonPage.tableUtils().checkIfExists(false);
-      commonPage.emptyState().checkIfExists(true);
-    });
-
-    it("Should search existing client scope by assigned type", () => {
-      commonPage
-        .tableToolbarUtils()
-        .selectSearchType(Filter.Name, Filter.AssignedType)
-        .selectSecondarySearchType(FilterAssignedType.Default);
-      commonPage
-        .tableUtils()
-        .checkRowItemExists(FilterAssignedType.Default)
-        .checkRowItemExists(FilterAssignedType.Optional, false);
-      commonPage
-        .tableToolbarUtils()
-        .selectSecondarySearchType(FilterAssignedType.Optional);
-      commonPage
-        .tableUtils()
-        .checkRowItemExists(FilterAssignedType.Default, false)
-        .checkRowItemExists(FilterAssignedType.Optional);
-      commonPage
-        .tableToolbarUtils()
-        .selectSecondarySearchType(FilterAssignedType.AllTypes);
-      commonPage
-        .tableUtils()
-        .checkRowItemExists(FilterAssignedType.Default)
-        .checkRowItemExists(FilterAssignedType.Optional);
-    });
-
-    const newItemsWithExpectedAssignedTypes = [
-      [clientScopeNameOptionalType, FilterAssignedType.Optional],
-      [clientScopeNameDefaultType, FilterAssignedType.Default],
-    ];
-    newItemsWithExpectedAssignedTypes.forEach(($type) => {
-      const [itemName, assignedType] = $type;
-      it(`Should add client scope ${itemName} with ${assignedType} assigned type`, () => {
-        commonPage.tableToolbarUtils().addClientScope();
-        commonPage
-          .modalUtils()
-          .checkModalTitle("Add client scopes to " + clientId);
-        commonPage.tableUtils().selectRowItemCheckbox(itemName);
-        commonPage.modalUtils().confirmModalWithItem(assignedType);
-        commonPage.masthead().checkNotificationMessage("Scope mapping updated");
-        commonPage.tableToolbarUtils().searchItem(itemName, false);
-        commonPage
-          .tableUtils()
-          .checkRowItemExists(itemName)
-          .checkRowItemExists(assignedType);
-      });
-    });
-
-    const expectedItemAssignedTypes = [
-      FilterAssignedType.Optional,
-      FilterAssignedType.Default,
-    ];
-    expectedItemAssignedTypes.forEach(($assignedType) => {
-      const itemName = clientScopeName + 0;
-      it(`Should change item ${itemName} AssignedType to ${$assignedType} from search bar`, () => {
-        commonPage.tableToolbarUtils().searchItem(itemName, false);
-        commonPage.tableUtils().selectRowItemCheckbox(itemName);
-        commonPage.tableToolbarUtils().changeTypeTo($assignedType);
-        commonPage.masthead().checkNotificationMessage("Scope mapping updated");
-        commonPage.tableToolbarUtils().searchItem(itemName, false);
-        commonPage.tableUtils().checkRowItemExists($assignedType);
-      });
-    });
-
-    it("Should show items on next page are more than 11", () => {
-      commonPage.sidebar().waitForPageLoad();
-      commonPage.tableToolbarUtils().clickNextPageButton();
-      commonPage.tableUtils().checkRowItemsGreaterThan(1);
-    });
-
-    it("Should remove client scope from item bar", () => {
-      const itemName = clientScopeName + 0;
-      commonPage.tableToolbarUtils().searchItem(itemName, false);
-      commonPage.tableUtils().selectRowItemAction(itemName, "Remove");
-      commonPage.modalUtils().confirmModal();
-      commonPage.masthead().checkNotificationMessage(msgScopeMappingRemoved);
-      commonPage.tableToolbarUtils().searchItem(itemName, false);
-      listingPage.assertNoResults();
-    });
-
-    it("Should remove multiple client scopes from search bar", () => {
-      const itemName1 = clientScopeName + 1;
-      const itemName2 = clientScopeName + 2;
-      cy.intercept(`/admin/realms/${realmName}/client-scopes`).as("load");
-      commonPage.tableToolbarUtils().clickSearchButton();
-      cy.wait("@load");
-      cy.wait(1000);
-      commonPage.tableToolbarUtils().checkActionItemIsEnabled("Remove", false);
-      commonPage.tableToolbarUtils().searchItem(clientScopeName, false);
-      commonPage
-        .tableUtils()
-        .selectRowItemCheckbox(itemName1)
-        .selectRowItemCheckbox(itemName2);
-      cy.intercept(`/admin/realms/${realmName}/client-scopes`).as("load");
-      commonPage.tableToolbarUtils().clickSearchButton();
-      cy.wait("@load");
-      cy.wait(1000);
-      commonPage.tableToolbarUtils().clickActionItem("Remove");
-      commonPage.masthead().checkNotificationMessage(msgScopeMappingRemoved);
-      commonPage.tableToolbarUtils().searchItem(clientScopeName, false);
-      commonPage
-        .tableUtils()
-        .checkRowItemExists(itemName1, false)
-        .checkRowItemExists(itemName2, false);
-      commonPage.tableToolbarUtils().clickSearchButton();
-    });
-
-    it("Should show initial items after filtering", () => {
-      commonPage
-        .tableToolbarUtils()
-        .selectSearchType(Filter.Name, Filter.AssignedType)
-        .selectSecondarySearchType(FilterAssignedType.Optional)
-        .selectSearchType(Filter.AssignedType, Filter.Name);
-      commonPage
-        .tableUtils()
-        .checkRowItemExists(FilterAssignedType.Default, false)
-        .checkRowItemExists(FilterAssignedType.Optional);
-    });
-  });
-
-  describe("Client scopes evaluate subtab", () => {
-    const clientName = "testClient";
-
-    beforeEach(() => {
-      loginPage.logIn();
-      keycloakBefore();
-      commonPage.sidebar().goToRealm(realmName);
-      commonPage.sidebar().goToClients();
-    });
-
-    before(async () => {
-      await adminClient.inRealm(realmName, () =>
-        adminClient.createClient({
-          protocol: "openid-connect",
-          clientId: clientName,
-          publicClient: false,
-        }),
-      );
-      await adminClient.inRealm(realmName, () =>
-        adminClient.createUser({
-          username: "admin-a",
-          enabled: true,
-        }),
-      );
-    });
-
-    after(async () => {
-      await adminClient.inRealm(realmName, () =>
-        adminClient.deleteClient(clientName),
-      );
-    });
-
-    it("check effective protocol mappers list is not empty and find effective protocol mapper locale", () => {
-      commonPage.tableToolbarUtils().searchItem(clientName);
-      commonPage.tableUtils().clickRowItemLink(clientName);
-
-      clientDetailsPage.goToClientScopesEvaluateTab();
-
-      cy.findByTestId("effective-protocol-mappers")
-        .find("tr")
-        .should("have.length.gt", 0);
-    });
-
-    it("check role scope mappings list list is not empty and find role scope mapping admin", () => {
-      commonPage.tableToolbarUtils().searchItem(clientName);
-      commonPage.tableUtils().clickRowItemLink(clientName);
-
-      clientDetailsPage.goToClientScopesEvaluateTab();
-      clientDetailsPage.goToClientScopesEvaluateEffectiveRoleScopeMappingsTab();
-
-      cy.findByTestId("effective-role-scope-mappings")
-        .find("tr")
-        .should("have.length.gt", 0);
-    });
-
-    it("check generated id token and user info", () => {
-      commonPage.tableToolbarUtils().searchItem(clientName);
-      commonPage.tableUtils().clickRowItemLink(clientName);
-
-      clientDetailsPage.goToClientScopesEvaluateTab();
-      cy.get("div#generatedAccessToken").contains("No generated access token");
-
-      clientDetailsPage.goToClientScopesEvaluateGeneratedIdTokenTab();
-      cy.get("div#generatedIdToken").contains("No generated id token");
-
-      clientDetailsPage.goToClientScopesEvaluateGeneratedUserInfoTab();
-      cy.get("div#generatedUserInfo").contains("No generated user info");
-
-      cy.get("[data-testid='user'] input").type("admin-a");
-      cy.get(".pf-v5-c-menu__item-text").click();
-
-      clientDetailsPage.goToClientScopesEvaluateGeneratedAccessTokenTab();
-      cy.get("div#generatedAccessToken").contains(
-        '"preferred_username": "admin-a"',
-      );
-      cy.get("div#generatedAccessToken").contains('"scope": "');
-
-      clientDetailsPage.goToClientScopesEvaluateGeneratedIdTokenTab();
-      cy.get("div#generatedIdToken").contains(
-        '"preferred_username": "admin-a"',
-      );
-
-      clientDetailsPage.goToClientScopesEvaluateGeneratedUserInfoTab();
-      cy.get("div#generatedIdToken").contains(
-        '"preferred_username": "admin-a"',
-      );
-      cy.get("div#generatedIdToken").contains('"sid"');
-    });
-  });
 
   describe("Client creation", () => {
     beforeEach(() => {
@@ -476,77 +159,6 @@ describe("Clients test", () => {
       commonPage.tableUtils().checkRowItemExists(itemId, false);
     });
 
-    it("Initial access token can't be created with 0 days and count", () => {
-      const initialAccessTokenTab = new InitialAccessTokenTab();
-      initialAccessTokenTab
-        .goToInitialAccessTokenTab()
-        .shouldBeEmpty()
-        .goToCreateFromEmptyList()
-        .fillNewTokenData(0, 0)
-        .checkExpirationGreaterThanZeroError()
-        .checkCountValue(1)
-        .checkSaveButtonIsDisabled();
-    });
-
-    it("Initial access token", () => {
-      const initialAccessTokenTab = new InitialAccessTokenTab();
-      initialAccessTokenTab
-        .goToInitialAccessTokenTab()
-        .shouldBeEmpty()
-        .goToCreateFromEmptyList()
-        .fillNewTokenData(1, 3)
-        .save();
-
-      commonPage
-        .modalUtils()
-        .checkModalTitle("Initial access token details")
-        .closeModal();
-
-      commonPage
-        .masthead()
-        .checkNotificationMessage("New initial access token has been created");
-
-      initialAccessTokenTab.shouldNotBeEmpty();
-
-      commonPage.tableToolbarUtils().searchItem("John Doe", false);
-      commonPage.emptyState().checkIfExists(true);
-      commonPage.tableToolbarUtils().searchItem("", false);
-
-      initialAccessTokenTab.getFirstId((id) => {
-        commonPage
-          .tableUtils()
-          .checkRowItemValueByItemName(id, 4, "4")
-          .checkRowItemValueByItemName(id, 5, "4")
-          .checkRowItemExists(id);
-      });
-
-      commonPage.tableToolbarUtils().clickPrimaryButton("Create");
-      initialAccessTokenTab.fillNewTokenData(1, 3).save();
-
-      commonPage.modalUtils().closeModal();
-
-      initialAccessTokenTab.getFirstId((id) => {
-        commonPage.tableUtils().selectRowItemAction(id, "Delete");
-        commonPage.sidebar().waitForPageLoad();
-        commonPage
-          .modalUtils()
-          .checkModalTitle("Delete initial access token?")
-          .confirmModal();
-      });
-
-      commonPage
-        .masthead()
-        .checkNotificationMessage("Initial access token deleted successfully");
-      initialAccessTokenTab.shouldNotBeEmpty();
-
-      initialAccessTokenTab.getFirstId((id) => {
-        commonPage.tableUtils().selectRowItemAction(id, "Delete");
-        commonPage.sidebar().waitForPageLoad();
-        commonPage.modalUtils().confirmModal();
-      });
-      initialAccessTokenTab.shouldBeEmpty();
-    });
-
     it("Should fail to create imported client with empty ID", () => {
       commonPage.sidebar().goToClients();
       cy.findByTestId("importClient").click();
@@ -599,365 +211,9 @@ describe("Clients test", () => {
     });
   });
 
-  describe("Roles tab test", () => {
-    const rolesTab = new ClientRolesTab();
-    const client = "client_" + crypto.randomUUID();
-    const createRealmRoleName = `create-realm-${crypto.randomUUID()}`;
-
-    before(async () => {
-      await adminClient.inRealm(realmName, () =>
-        adminClient.createClient({
-          clientId: client,
-          protocol: "openid-connect",
-          publicClient: false,
-        }),
-      );
-      await adminClient.inRealm(realmName, () =>
-        adminClient.createRealmRole({
-          name: createRealmRoleName,
-        }),
-      );
-    });
-
-    beforeEach(() => {
-      loginPage.logIn();
-      keycloakBefore();
-      commonPage.sidebar().goToRealm(realmName);
-      commonPage.sidebar().goToClients();
-      commonPage.tableToolbarUtils().searchItem(client);
-      commonPage.tableUtils().clickRowItemLink(client);
-      rolesTab.goToRolesTab();
-    });
-
-    after(async () => {
-      await adminClient.inRealm(realmName, () =>
-        adminClient.deleteClient(client),
-      );
-      await adminClient.inRealm(realmName, () =>
-        adminClient.deleteRealmRole(createRealmRoleName),
-      );
-    });
-
-    it("Should fail to create client role with empty name", () => {
-      rolesTab.goToCreateRoleFromEmptyState();
-      createRealmRolePage.fillRealmRoleData("").save();
-      createRealmRolePage.checkRealmRoleNameRequiredMessage();
-    });
-
-    it("Should create client role", () => {
-      rolesTab.goToCreateRoleFromEmptyState();
-      createRealmRolePage.fillRealmRoleData(itemId).save();
-      commonPage.masthead().checkNotificationMessage("Role created", true);
-    });
-
-    it("Should update client role description", () => {
-      const updateDescription = "updated description";
-      commonPage.tableToolbarUtils().searchItem(itemId, false);
-      commonPage.tableUtils().clickRowItemLink(itemId);
-      createRealmRolePage.updateDescription(updateDescription).save();
-      commonPage
-        .masthead()
-        .checkNotificationMessage("The role has been saved", true);
-      createRealmRolePage.checkDescription(updateDescription);
-    });
-
-    it("Should add attribute to client role", () => {
-      commonPage.tableUtils().clickRowItemLink(itemId);
-      rolesTab.goToAttributesTab();
-      attributesTab
-        .addAttribute("crud_attribute_key", "crud_attribute_value")
-        .save();
-      attributesTab.assertRowItemsEqualTo(1);
-      commonPage
-        .masthead()
-        .checkNotificationMessage("The role has been saved", true);
-    });
-
-    it("Should delete attribute from client role", () => {
-      commonPage.tableUtils().clickRowItemLink(itemId);
-      rolesTab.goToAttributesTab();
-      attributesTab.deleteAttribute(0);
-      attributesTab.assertEmpty();
-      commonPage
-        .masthead()
-        .checkNotificationMessage("The role has been saved", true);
-    });
-
-    it("Should create client role to be deleted", () => {
-      rolesTab.goToCreateRoleFromToolbar();
-      createRealmRolePage.fillRealmRoleData("client_role_to_be_deleted").save();
-      commonPage.masthead().checkNotificationMessage("Role created", true);
-    });
-
-    it("Should fail to create duplicate client role", () => {
-      rolesTab.goToCreateRoleFromToolbar();
-      createRealmRolePage.fillRealmRoleData(itemId).save();
-      commonPage
-        .masthead()
-        .checkNotificationMessage(
-          `Could not create role: Role with name ${itemId} already exists`,
-          true,
-        );
-    });
-
-    it("Should search existing client role", () => {
-      commonPage.tableToolbarUtils().searchItem(itemId, false);
-      commonPage.tableUtils().checkRowItemExists(itemId);
-    });
-
-    it("Should search non-existing role test", () => {
-      commonPage.tableToolbarUtils().searchItem("role_DNE", false);
-      commonPage.emptyState().checkIfExists(true);
-    });
-
-    it("roles empty search test", () => {
-      commonPage.tableToolbarUtils().searchItem("", false);
-      commonPage.tableUtils().checkIfExists(true);
-    });
-
-    it("Add associated roles test", () => {
-      commonPage.tableToolbarUtils().searchItem(itemId, false);
-      commonPage.tableUtils().clickRowItemLink(itemId);
-
-      // Add associated realm role
-      associatedRolesPage.addAssociatedRealmRole(createRealmRoleName);
-      commonPage
-        .masthead()
-        .checkNotificationMessage("Associated roles have been added", true);
-
-      // Add associated client role
-      associatedRolesPage.addAssociatedRoleFromSearchBar(
-        "manage-account",
-        true,
-      );
-      commonPage
-        .masthead()
-        .checkNotificationMessage("Associated roles have been added", true);
-
-      rolesTab.goToAssociatedRolesTab();
-
-      // Add associated client role
-      associatedRolesPage.addAssociatedRoleFromSearchBar(
-        "manage-consent",
-        true,
-      );
-      commonPage
-        .masthead()
-        .checkNotificationMessage("Associated roles have been added", true);
-    });
-
-    it("Should hide inherited roles test", () => {
-      commonPage.tableToolbarUtils().searchItem(itemId, false);
-      commonPage.tableUtils().clickRowItemLink(itemId);
-      rolesTab.goToAssociatedRolesTab().hideInheritedRoles();
-    });
-
-    it("Should delete associated roles test", () => {
-      commonPage.tableToolbarUtils().searchItem(itemId, false);
-      commonPage.tableUtils().clickRowItemLink(itemId);
-      rolesTab.goToAssociatedRolesTab();
-      commonPage
-        .tableUtils()
-        .selectRowItemAction(createRealmRoleName, "Unassign");
-      commonPage.sidebar().waitForPageLoad();
-      commonPage.modalUtils().checkModalTitle("Remove role?").confirmModal();
-      commonPage.sidebar().waitForPageLoad();
-
-      commonPage
-        .masthead()
-        .checkNotificationMessage("Scope mapping successfully removed", true);
-
-      commonPage.tableUtils().selectRowItemAction("manage-consent", "Unassign");
-      commonPage.sidebar().waitForPageLoad();
-      commonPage.modalUtils().checkModalTitle("Remove role?").confirmModal();
-    });
-
-    it("Should delete associated role from search bar test", () => {
-      commonPage.tableToolbarUtils().searchItem(itemId, false);
-      commonPage.tableUtils().clickRowItemLink(itemId);
-      commonPage.sidebar().waitForPageLoad();
-      rolesTab.goToAssociatedRolesTab();
-
-      cy.get("td")
-        .contains("manage-account")
-        .parent()
-        .within(() => {
-          cy.get("input").click();
-        });
-
-      associatedRolesPage.removeAssociatedRoles();
-
-      commonPage.sidebar().waitForPageLoad();
-      commonPage.modalUtils().checkModalTitle("Remove role?").confirmModal();
-      commonPage.sidebar().waitForPageLoad();
-
-      commonPage
-        .masthead()
-        .checkNotificationMessage("Scope mapping successfully removed", true);
-    });
-
-    it("Should delete client role test", () => {
-      commonPage.tableUtils().selectRowItemAction(itemId, "Delete");
-      commonPage.sidebar().waitForPageLoad();
-      commonPage.modalUtils().checkModalTitle("Delete role?").confirmModal();
-    });
-
-    it("Should delete client role from role details test", () => {
-      commonPage
-        .tableToolbarUtils()
-        .searchItem("client_role_to_be_deleted", false);
-      commonPage.tableUtils().clickRowItemLink("client_role_to_be_deleted");
-      createRealmRolePage.clickActionMenu("Delete this role");
-      commonPage.modalUtils().confirmModal();
-      commonPage
-        .masthead()
-        .checkNotificationMessage("The role has been deleted", true);
-    });
-  });
-
-  describe("Advanced tab test", () => {
-    const advancedTab = new AdvancedTab();
-    let client: string;
-
-    beforeEach(() => {
-      loginPage.logIn();
-      keycloakBefore();
-      commonPage.sidebar().goToRealm(realmName);
-      commonPage.sidebar().goToClients();
-      client = "client_" + crypto.randomUUID();
-      commonPage.tableToolbarUtils().createClient();
-      createClientPage
-        .selectClientType("OpenID Connect")
-        .fillClientData(client)
-        .continue();
-
-      commonPage.sidebar().waitForPageLoad();
-
-      createClientPage.continue().save();
-      commonPage
-        .masthead()
-        .checkNotificationMessage("Client created successfully");
-      clientDetailsPage.goToAdvancedTab();
-    });
-
-    afterEach(() =>
-      adminClient.inRealm(realmName, () => adminClient.deleteClient(client)),
-    );
-
-    it("Clustering", () => {
-      advancedTab.expandClusterNode();
-
-      advancedTab.checkEmptyClusterNode();
-
-      advancedTab.registerNodeManually().fillHost("localhost").saveHost();
-      advancedTab.checkTestClusterAvailability(true);
-      commonPage.masthead().checkNotificationMessage("Node successfully added");
-      advancedTab.deleteClusterNode();
-      commonPage.modalUtils().confirmModal();
-      commonPage
-        .masthead()
-        .checkNotificationMessage("Node successfully removed");
-      advancedTab.checkEmptyClusterNode();
-    });
-
-    it("Fine grain OpenID connect configuration", () => {
-      const algorithm = "ES384";
-      advancedTab
-        .selectAccessTokenSignatureAlgorithm(algorithm)
-        .saveFineGrain();
-
-      advancedTab
-        .selectAccessTokenSignatureAlgorithm("HS384")
-        .revertFineGrain();
-      advancedTab.checkAccessTokenSignatureAlgorithm(algorithm);
-    });
-
-    it("OIDC Compatibility Modes configuration", () => {
-      advancedTab.clickAllCompatibilitySwitch();
-      advancedTab.saveCompatibility();
-      advancedTab.jumpToCompatability();
-      advancedTab.clickExcludeSessionStateSwitch();
-      advancedTab.clickUseRefreshTokenForClientCredentialsGrantSwitch();
-      advancedTab.revertCompatibility();
-    });
-
-    it("Client Offline Session Max", () => {
-      configureOfflineSessionMaxInRealmSettings(true);
-
-      cy.findByTestId("token-lifespan-clientOfflineSessionMax").should("exist");
-
-      configureOfflineSessionMaxInRealmSettings(false);
-
-      cy.findByTestId("token-lifespan-clientOfflineSessionMax").should(
-        "not.exist",
-      );
-
-      function configureOfflineSessionMaxInRealmSettings(enabled: boolean) {
-        commonPage.sidebar().goToRealmSettings();
-        realmSettings.goToSessionsTab();
-        realmSettings.setOfflineSessionMaxSwitch(enabled);
-        realmSettings.saveSessions();
-
-        commonPage.sidebar().goToClients();
-        commonPage.tableToolbarUtils().searchItem(client);
-        commonPage.tableUtils().clickRowItemLink(client);
-        clientDetailsPage.goToAdvancedTab();
-      }
-    });
-
-    it("Advanced settings", () => {
-      advancedTab.jumpToAdvanced();
-
-      advancedTab.clickAdvancedSwitches();
-      advancedTab.jumpToAdvanced();
-      advancedTab.selectKeyForCodeExchangeInput("S256");
-
-      advancedTab.saveAdvanced();
-      advancedTab.jumpToAdvanced();
-      advancedTab.checkAdvancedSwitchesOn();
-      advancedTab.checkKeyForCodeExchangeInput("S256");
-
-      advancedTab.selectKeyForCodeExchangeInput("plain");
-      advancedTab.checkKeyForCodeExchangeInput("plain");
-
-      advancedTab.jumpToAdvanced();
-      advancedTab.clickAdvancedSwitches();
-
-      advancedTab.revertAdvanced();
-      advancedTab.jumpToAdvanced();
-      advancedTab.checkKeyForCodeExchangeInput("S256");
-      //uncomment when revert button reverts all switches
-      //and ACR to LoA Mapping + Default ACR Values
-      //advancedTab.checkAdvancedSwitchesOn();
-    });
-
-    it("Authentication flow override", () => {
-      advancedTab.jumpToAuthFlow();
-      advancedTab.selectBrowserFlowInput("browser");
-      advancedTab.selectDirectGrantInput("docker auth");
-      advancedTab.checkBrowserFlowInput("browser");
-      advancedTab.checkDirectGrantInput("docker auth");
-
-      advancedTab.revertAuthFlowOverride();
-      advancedTab.jumpToAuthFlow();
-      advancedTab.checkBrowserFlowInput("Choose...");
-      advancedTab.checkDirectGrantInput("Choose...");
-      advancedTab.selectBrowserFlowInput("browser");
-      advancedTab.selectDirectGrantInput("docker auth");
-
-      advancedTab.saveAuthFlowOverride();
-      advancedTab.selectBrowserFlowInput("first broker login");
-      advancedTab.selectDirectGrantInput("first broker login");
-      advancedTab.revertAuthFlowOverride();
-      //revert doesn't work after saving.
-      //advancedTab.CheckBrowserFlowInput("browser");
-      //advancedTab.CheckDirectGrantInput("docker auth");
-    });
-  });
-
   describe("Service account tab test", () => {
     const serviceAccountTab = new RoleMappingTab("user");
-    const serviceAccountName = "service-account-client";
+    const serviceAccountName = `service-account-client${crypto.randomUUID()}`;
     const createRealmRoleName = `create-realm-${crypto.randomUUID()}`;
     const createRealmRoleType = `roles`;
 
@@ -1344,12 +600,6 @@ describe("Clients test", () => {
       commonPage.tableUtils().selectRowItemAction(clientId, "Delete");
       cy.checkA11y();
       cy.findAllByTestId("confirm").click();
-    });
-
-    it("Check a11y violations on initial access token", () => {
-      const initialAccessTokenTab = new InitialAccessTokenTab();
-      initialAccessTokenTab.goToInitialAccessTokenTab();
-      cy.checkA11y();
     });
 
     it("Check a11y violations on client registration/ anonymous access policies tab", () => {
